refactor cmd/frontman in order to make more code testable
