+++ pkg/winui/winui_fallback.go
@@ -1,0 +1,12 @@
+// +build !windows
+
+package winui
+
+import (
+	"github.com/cloudradar-monitoring/frontman"
+)
+
+// WindowsShowSettingsUI stub exists only for cross-platform compiling on platforms that don't implement Windows UI.
+func WindowsShowSettingsUI(_ *frontman.Frontman, _ bool) {
+
+}
