+++ pkg/winui/winui_service.go
@@ -1,6 +1,6 @@
 // +build windows
 
-package main
+package winui
 
 import (
 	"context"
